+++ src/app/app.js
@@ -1,0 +1,5 @@
+import loadPokemons from './utils/loadpokemons.js';
+
+window.onload = () => {
+  loadPokemons();
+};
