+++ src/template.html
@@ -46,8 +46,6 @@
 
       <div class="container row row-cols-1 row-cols-sm-2 row-cols-md-3 row-cols-lg-4 g-5 mx-auto"></div>
 
-      <div class="container2 d-flex"></div>
-
 
       <footer class="d-flex flex-wrap justify-content-between align-items-center py-3 my-4 border-top">
         <div class="col-md-4 d-flex align-items-center">
